Adding support for all known extension types to this date and improved code style
It's not possible using the manifest cache because it doesn't cache the entire manifest, just bits and pieces. Using the manifest cache would require changes in the Joomla! Platform (JInstaller) and a method to regenerate the manifest cache for all installed extensions automatically. This is extreme overkill and way out of the scope of our work.